ref #29216 - Exception Classes, process get and delete with 404 and 403

+++ src/Zmsapi/ProcessDelete.php
@@ -21,10 +21,18 @@ class ProcessDelete extends BaseController
     {
         $query = new Query();
         $message = Response\Message::create(Render::$request);
-        $process = $query->readEntity($itemId, $authKey);
-        $query->deleteEntity($itemId, $authKey);
+        $authKeyByProcessId = $query->readAuthKeyByProcessId($itemId);
 
-        $message->data = $process;
+        if (null === $authKeyByProcessId) {
+            throw new Exception\Process\ProcessNotFound();
+        } elseif ($authKeyByProcessId != $authKey) {
+            throw new Exception\Process\AuthKeyMatchFailed();
+        } else {
+            $process = $query->readEntity($itemId, $authKey);
+            $query->deleteEntity($itemId, $authKey);
+            $message->data = $process;
+        }
+
         Render::lastModified(time(), '0');
         Render::json($message->setUpdatedMetaData(), $message->getStatuscode());
     }
